requirements.txt
anthropic==0.75.0
google-cloud-aiplatform==1.98.0
google-cloud-storage>=2.16.0,<3.0.0
google-cloud-resource-manager>=1.3.3,<3.0.0
google-cloud-bigquery>=3.37.0,<4.0.0
google-cloud-core>=2.4.1,<3.0.0
google-resumable-media>=2.7.0,<3.0.0
grpc-google-iam-v1>=0.12.3,<1.0.0
google-crc32c>=1.7.1,<2.0.0
protobuf>=5.26.1,<6.0.0
google-genai==1.21.1
fastapi==0.124.4
firebase-admin
httpx==0.28.1
langchain_anthropic==1.3.0
langchain_core==1.2.2
langchain_openai==1.1.4
numpy==2.3.5
openai==2.13.0
openpyxl
pandas==2.3.3
Pillow==12.0.0
pydantic==2.12.5
pydantic_settings==2.12.0
pymongo==4.15.5
PyMySQL==1.1.1
python-dotenv==1.2.1
python-multipart==0.0.20
reportlab==4.4.5
Requests==2.32.5
SQLAlchemy==2.0.45
starlette==0.50.0
tldextract==5.3.0
uvicorn
sentence-transformers
scikit-learn
langchain_huggingface
streamlit
tenacity
python-docx
